feat: orderLineItem과 menuProduct 의존성 제거

+++ src/main/java/kitchenpos/menu/application/dto/MenuProductResponse.java
@@ -22,7 +22,7 @@ public class MenuProductResponse {
                 menuProduct.getSeq(),
                 menuProduct.getMenuId(),
                 menuProduct.getProductName(),
-                menuProduct.getProductPrice(),
+                menuProduct.getProductPrice().longValue(),
                 menuProduct.getQuantity());
     }
